Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15493 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/pluginSrc/NDFileJPEG.h
@@ -53,6 +53,6 @@ private:
     JOCTET jpegBuffer[JPEG_BUF_SIZE];
     jpegDestMgr destMgr;
 };
-#define NUM_NDFILE_JPEG_PARAMS (&LAST_NDFILE_JPEG_PARAM - &FIRST_NDFILE_JPEG_PARAM + 1)
+#define NUM_NDFILE_JPEG_PARAMS ((int)(&LAST_NDFILE_JPEG_PARAM - &FIRST_NDFILE_JPEG_PARAM + 1))
 
 #endif
